fix: comprehensive platform bug fixes across 19 files
Critical execution fixes:
- stop-loss orders: TimeInForce.OPG → DAY (OPG never fills during market hours)
- stop-loss orders: wrap submit_order with DRY_RUN (was bypassing test mode)
- CashManager: num_strategies=4 → len(CANONICAL_STRATEGY_SPECS)=7
- SELL trade_record: used requested qty instead of actual_filled_qty
- Add reconciliation_status/discrepancies/strategies_cache to __init__ defaults
- Store strategies as strategies_cache after initialize_strategies()

Database fixes:
- Remove 7 duplicate method definitions (Python silently used second copy)
- drawdown_pct sign inverted: (portfolio-peak)/peak → (peak-portfolio)/peak
- avg_loss_pct: abs(mean(losses)) → mean(abs(losses)) for correct Kelly criterion

Data pipeline fixes:
- VWAP: cumulative sum → 20-day rolling (was growing unbounded across dates)
- ADX: guard zero denominator with .replace(0, nan) to prevent NaN propagation
- data_validator: and → or in staleness check (date-gap check was dead code)
- data_validator: early return after missing-column detection (prevented KeyError)
- data_validator: guard tz-aware vs naive datetime comparison

Risk management fixes:
- stop_loss_manager: if not entry → if entry is None (0.0 treated as missing)
- broker_reconciler: int(pos.qty) → float(pos.qty) (truncated fractional shares)
- drawdown_stop_manager: or pattern → is not None (intentional 0.0 was ignored)
- drawdown_stop_manager: guard fromisoformat(None) crash on null state fields
- dynamic_allocator: guard division by zero in rebalance_needed when target=0
- correlation_filter: fix missing check_correlation → should_filter_signal

Strategy fixes:
- All 4 strategies: LTCG hold 250 → 365 calendar days (firing 3+ months early)
- strategy_base: guard price=0 in calculate_position_size and get_return_pct
- MA crossover: guard sma50=0 division in gap_pct confidence formula
- News sentiment: reorder max_hold_days before hold_days (was dead elif branch)
- sync_broker_state: str(positions) → json.dumps (was storing Python repr)

Metrics fixes:
- performance_metrics: net_pnl double-deducted costs (pnl already net of costs)
- performance_metrics: annualized return linear → CAGR formula
- performance_metrics: Sharpe/Sortino use ddof=1 (sample) not ddof=0 (population)
- strategy_health_scorer: win_rate denominator = all trades → SELL trades only
- generate_daily_email: fmt_money negative sign placed correctly (-$X not $-X)

+++ src/risk/broker_reconciler.py
@@ -339,7 +339,7 @@ class BrokerReconciler:
             return {
                 "positions": {
                     pos.symbol: {
-                        "qty": int(pos.qty),
+                        "qty": float(pos.qty),
                         "avg_price": float(pos.avg_entry_price),
                         "market_value": float(pos.market_value),
                         "unrealized_pl": float(pos.unrealized_pl),
